Summary uptage Top book
Розбив код на декілька файлів для кращої читабельності.
Додав підвантаження на кнопку See more
